push more methods out of _tdb_handles.py
