+++ apis/python/src/tiledbsoma/io/ingest.py
@@ -1931,7 +1931,7 @@ def _write_matrix_to_denseNDArray(
 
 def _read_nonempty_domain(arr: SOMAArray) -> Any:
     try:
-        return arr._handle.non_empty_domain()
+        return arr.non_empty_domain()
     except (SOMAError, RuntimeError):
         # This means that we're open in write-only mode.
         # Reopen the array in read mode.
@@ -1940,7 +1940,7 @@ def _read_nonempty_domain(arr: SOMAArray) -> Any:
 
     cls = type(arr)
     with cls.open(arr.uri, "r", platform_config=None, context=arr.context) as readarr:
-        return readarr._handle.non_empty_domain()
+        return readarr.non_empty_domain()
 
 
 def _find_sparse_chunk_size(
